+++ src/PrunePolicyDecay.cc
@@ -1,4 +1,4 @@
-// Copyright © 2015 Richard Kettlewell.
+// Copyright © Richard Kettlewell.
 //
 // This program is free software: you can redistribute it and/or modify
 // it under the terms of the GNU General Public License as published by
@@ -88,7 +88,7 @@ public:
     // preserved.
     std::map<int, const Backup *> oldest;
     for(Backup *backup: onDevice) {
-      int age = Date::today() - Date(backup->time);
+      int age = Date::today("PRUNE") - Date(backup->time);
       // Keep backups that are young enough
       int a = age - decayStart;
       if(a <= 0)
@@ -112,7 +112,7 @@ public:
     // Now that we know what the oldest backup in each bucket is, we can prune
     // the rest.
     for(Backup *backup: onDevice) {
-      int age = Date::today() - Date(backup->time);
+      int age = Date::today("PRUNE") - Date(backup->time);
       // Keep backups that are young enough
       int a = age - decayStart;
       if(a <= 0 || age > decayLimit)